Split interceptors and split router into subrouters to allow per router interceptors
