+++ handlers/register.go
@@ -2,6 +2,7 @@
 package handlers
 
 import (
+	"github.com/companieshouse/chs.go/log"
 	"github.com/companieshouse/payments.api.ch.gov.uk/interceptors"
 	"net/http"
 
@@ -22,18 +23,27 @@ func Register(mainRouter *mux.Router, cfg config.Config) {
 	}
 
 	mainRouter.HandleFunc("/healthcheck", healthCheck).Methods("GET").Name("get-healthcheck")
+	mainRouter.Use(log.Handler)
 
-	// Create subrouters. All routes except /callback need auth middleware, so router needs to be split up.
-	paymentsRouter := mainRouter.PathPrefix("/payments").Subrouter()
-	paymentsRouter.HandleFunc("", p.CreatePaymentSession).Methods("POST").Name("create-payment")
-	paymentsRouter.HandleFunc("/{payment_id}", p.GetPaymentSession).Methods("GET").Name("get-payment")
+	// Create subrouters. All routes except /callback need auth middleware, so router needs to be split up. This allows
+	// per-subrouter middleware.
 
+	// create-payment endpoint should not be intercepted by the paymentauth interceptor, so needs to be it's own subrouter
+	rootPaymentRouter := mainRouter.PathPrefix("/payments").Subrouter()
+	rootPaymentRouter.HandleFunc("", p.CreatePaymentSession).Methods("POST").Name("create-payment")
+
+	// get-payment endpoint needs payment and user auth, so needs to be it's own subrouter
+	getPaymentRouter := rootPaymentRouter.PathPrefix("/{payment_id}").Subrouter()
+	getPaymentRouter.HandleFunc("", p.GetPaymentSessionFromRequest).Methods("GET").Name("get-payment")
+
+	// All private endpoints need  payment and user auth, so needs to be it's own subrouter
 	privateRouter := mainRouter.PathPrefix("/private").Subrouter()
 	privateRouter.HandleFunc("/payments/{payment_id}", p.PatchPaymentSession).Methods("PATCH").Name("patch-payment")
 	privateRouter.HandleFunc("/payments/{payment_id}/external-journey", p.CreateExternalPaymentJourney).Methods("POST").Name("create-external-payment-journey")
 
 	// Set middleware for subrouters
-	paymentsRouter.Use(interceptors.UserAuthenticationInterceptor, interceptors.PaymentAuthenticationInterceptor)
+	rootPaymentRouter.Use(interceptors.UserAuthenticationInterceptor)
+	getPaymentRouter.Use(interceptors.UserAuthenticationInterceptor, interceptors.PaymentAuthenticationInterceptor)
 	privateRouter.Use(interceptors.UserAuthenticationInterceptor, interceptors.PaymentAuthenticationInterceptor)
 }
