+++ service/payment_test.go
@@ -25,7 +25,7 @@ var defaultCost = models.CostResource{
 	ClassOfPayment:          []string{"class"},
 	Description:             "desc",
 	DescriptionIdentifier:   "identifier",
-	Links:                   models.Links{Self: "self"},
+	Links: models.Links{Self: "self"},
 }
 
 var defaultCostArray = []models.CostResource{
@@ -254,7 +254,7 @@ func TestUnitGetPayment(t *testing.T) {
 		req, err := http.NewRequest("Get", "", nil)
 		So(err, ShouldBeNil)
 		w := httptest.NewRecorder()
-		mockPaymentService.GetPaymentSession(w, req)
+		mockPaymentService.GetPaymentSessionFromRequest(w, req)
 		So(w.Code, ShouldEqual, 400)
 	})
 
@@ -267,7 +267,7 @@ func TestUnitGetPayment(t *testing.T) {
 		req = mux.SetURLVars(req, map[string]string{"payment_id": "invalid"})
 		So(err, ShouldBeNil)
 		w := httptest.NewRecorder()
-		mockPaymentService.GetPaymentSession(w, req)
+		mockPaymentService.GetPaymentSessionFromRequest(w, req)
 		So(w.Code, ShouldEqual, 403)
 	})
 
@@ -280,7 +280,7 @@ func TestUnitGetPayment(t *testing.T) {
 		req = mux.SetURLVars(req, map[string]string{"payment_id": "1234"})
 		So(err, ShouldBeNil)
 		w := httptest.NewRecorder()
-		mockPaymentService.GetPaymentSession(w, req)
+		mockPaymentService.GetPaymentSessionFromRequest(w, req)
 		So(w.Code, ShouldEqual, 500)
 	})
 
@@ -293,7 +293,7 @@ func TestUnitGetPayment(t *testing.T) {
 		req = mux.SetURLVars(req, map[string]string{"payment_id": "1234"})
 		So(err, ShouldBeNil)
 		w := httptest.NewRecorder()
-		mockPaymentService.GetPaymentSession(w, req)
+		mockPaymentService.GetPaymentSessionFromRequest(w, req)
 		So(w.Code, ShouldEqual, 400)
 	})
 
@@ -318,7 +318,7 @@ func TestUnitGetPayment(t *testing.T) {
 
 		httpmock.RegisterResponder("GET", "http://dummy-resource", jsonResponse)
 		w := httptest.NewRecorder()
-		mockPaymentService.GetPaymentSession(w, req)
+		mockPaymentService.GetPaymentSessionFromRequest(w, req)
 		So(w.Code, ShouldEqual, 500)
 	})
 
@@ -338,7 +338,7 @@ func TestUnitGetPayment(t *testing.T) {
 		costArray[0].Amount = "99"
 		jsonResponse, _ := httpmock.NewJsonResponder(200, costArray)
 		httpmock.RegisterResponder("GET", "http://dummy-resource", jsonResponse)
-		mockPaymentService.GetPaymentSession(w, req)
+		mockPaymentService.GetPaymentSessionFromRequest(w, req)
 		So(w.Code, ShouldEqual, 403)
 	})
 
@@ -357,7 +357,7 @@ func TestUnitGetPayment(t *testing.T) {
 		costArray := []models.CostResource{defaultCost}
 		jsonResponse, _ := httpmock.NewJsonResponder(200, costArray)
 		httpmock.RegisterResponder("GET", "http://dummy-resource", jsonResponse)
-		mockPaymentService.GetPaymentSession(w, req)
+		mockPaymentService.GetPaymentSessionFromRequest(w, req)
 		So(w.Code, ShouldEqual, 200)
 	})
 
@@ -376,7 +376,7 @@ func TestUnitGetPayment(t *testing.T) {
 		costArray := []models.CostResource{defaultCost, defaultCost}
 		jsonResponse, _ := httpmock.NewJsonResponder(200, costArray)
 		httpmock.RegisterResponder("GET", "http://dummy-resource", jsonResponse)
-		mockPaymentService.GetPaymentSession(w, req)
+		mockPaymentService.GetPaymentSessionFromRequest(w, req)
 		So(w.Code, ShouldEqual, 200)
 	})
 }
@@ -538,7 +538,7 @@ func TestUnitValidateCosts(t *testing.T) {
 			ClassOfPayment:          []string{"class"},
 			Description:             "",
 			DescriptionIdentifier:   "identifier",
-			Links:                   models.Links{Self: "self"},
+			Links: models.Links{Self: "self"},
 		}}
 		So(validateCosts(&cost), ShouldNotBeNil)
 	})
@@ -549,7 +549,7 @@ func TestUnitValidateCosts(t *testing.T) {
 			ClassOfPayment:          []string{"class"},
 			Description:             "desc",
 			DescriptionIdentifier:   "identifier",
-			Links:                   models.Links{Self: "self"},
+			Links: models.Links{Self: "self"},
 		}}
 		So(validateCosts(&cost), ShouldBeNil)
 	})
@@ -561,7 +561,7 @@ func TestUnitValidateCosts(t *testing.T) {
 				ClassOfPayment:          []string{"class"},
 				Description:             "desc",
 				DescriptionIdentifier:   "identifier",
-				Links:                   models.Links{Self: "self"},
+				Links: models.Links{Self: "self"},
 			},
 			{
 				Amount:                  "20",
@@ -569,7 +569,7 @@ func TestUnitValidateCosts(t *testing.T) {
 				ClassOfPayment:          []string{"class"},
 				Description:             "",
 				DescriptionIdentifier:   "identifier",
-				Links:                   models.Links{Self: "self"},
+				Links: models.Links{Self: "self"},
 			},
 		}
 		So(validateCosts(&cost), ShouldNotBeNil)
